# Optimize move_player() and update_nearest_tile()