When moving player sprites, we called `update_nearest_tile()` twice when moving diagonally and did some math which we didn't need to do. Also added an early-return to `move_player()` in case the player isn't holding any d-pad buttons at all.

This takes `move_player()` from 130 cycles when idle to 78 cycles, when holding one direction from 1174 cycles to 487 cycles, and when holding two directions from 2196 cycles to 575 cycles.

Also improves `update_nearest_tile()` itself from using 730 cycles to 283 cycles.